Update Assets/Plugins/FMOD/RuntimeManager.cs
Fixed issues with the AttachInstanceToGameObject methods.
This mainly fixes issues when being called via the PlayOneShotAttached method.
AttachInstanceToGameObject  now directly sets the 3dAttribute while attaching to ensure an initial correct and valid 3d positioning is applied.
Before a 3d position would only by applied to the instance in the next update after the attach occured.
This can cause unexpected and unwanted results in case the target transform was destroyed between the attach and the update loop.

Example: A crate that plays a PlayOneShotAttached when it hits a surface but that also takes damage at the same time that can result in a destroy. The sound would play at the incorrect location as there was never an initial position set that matched the desired start position. This small change will fix this.