OnGUI Grabage Generation fix
Updated the DrawDebugOverlay to no longer cause Garbage allocation as long as the overlay isn't enabled.
Before: OnGUI would always be called resulting in the Unity OnGUI to be active no matter what, generating unwanted overhead and garbage.
Now: As long as the debugoverlay isn't active no garbage will be allocated. A safety is also added to avoid the overhead being applied in a non editor build by accident. Can be adjusted if desired.